pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src"]

[project]
name = "table-talks"
version = "1.0.0"
description = "Telegram bot for table-talk conversation cards by theme"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
]

[dependency-groups]
dev = [
    "basedpyright>=1.35",
    "pytest>=8.0",
    "ruff>=0.8",
]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.pyright]
pythonVersion = "3.10"
typeCheckingMode = "strict"
reportOptionalMemberAccess = true
reportOptionalSubscript = true
venvPath = "."
venv = ".venv"
